+++ CHANGELOG.rst
@@ -4,7 +4,9 @@ Version NEXT
 * models.FloatField are now automatically localized.
 * Support for Django 1.2 dropped.
 * Product model now has property ``can_be_added_to_cart`` which is checked before adding the product to cart
-
+* In cart_modifiers methods ``get_extra_cart_price_field`` and ``get_extra_cart_item_price_field``
+  accepts the additional field ``state`` to pass customized data from ``Cart.update()``
+  down to these methods.
 
 Version 0.1.2
 =============